Final_Backend/geniai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
pypdf==3.17.4
faiss-cpu==1.7.4
numpy==1.24.3
python-dotenv==1.0.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0